Use shared types for api hooks
