+++ src/hooks/api/events.ts
@@ -1,5 +1,4 @@
-import type { NextApiRequest } from 'next';
-import type { QueryClient, UseQueryOptions } from 'react-query';
+import type { UseQueryOptions } from 'react-query';
 
 import { fetchFromApi, isErrorObject } from '@/utils/fetchFromApi';
 import { formatDate } from '@/utils/formatDate';
@@ -9,18 +8,9 @@ import {
   useAuthenticatedQueries,
   useAuthenticatedQuery,
 } from './authenticated-query';
-
-type ServerSideArguments = {
-  req?: NextApiRequest;
-  queryClient?: QueryClient;
-};
-
-type Headers = Record<string, string>;
-
-type SortOptions = {
-  field: string;
-  order: string;
-};
+import type { Headers } from './types/Headers';
+import type { ServerSideArguments } from './types/ServerSideArguments';
+import type { SortOptions } from './types/SortOptions';
 
 type HeadersAndQueryData = {
   headers: Headers;
