+++ src/hooks/api/types/Headers.ts
@@ -1,0 +1,3 @@
+type Headers = Record<string, string>;
+
+export type { Headers };
