+++ src/hooks/api/types/ServerSideArguments.ts
@@ -1,0 +1,9 @@
+import type { NextApiRequest } from 'next';
+import type { QueryClient } from 'react-query';
+
+type ServerSideArguments = {
+  req?: NextApiRequest;
+  queryClient?: QueryClient;
+};
+
+export type { ServerSideArguments };
