+++ src/hooks/api/types/SortOptions.ts
@@ -1,0 +1,6 @@
+type SortOptions = {
+  field: string;
+  order: string;
+};
+
+export type { SortOptions };
